+++ packages/databricks-vscode/src/configuration/ui/ConfigurationDataProvider.ts
@@ -9,13 +9,13 @@ import {
 } from "vscode";
 
 import {ConnectionManager} from "../ConnectionManager";
-import {ConfigModel} from "../ConfigModel";
+import {ConfigModel} from "../models/ConfigModel";
 import {BaseComponent} from "./BaseComponent";
 import {ConfigurationTreeItem} from "./types";
 import {BundleTargetComponent} from "./BundleTargetComponent";
 import {AuthTypeComponent} from "./AuthTypeComponent";
 import {ClusterComponent} from "./ClusterComponent";
-import {WorkspaceComponent} from "./WorkspaceComponent";
+import {SyncDestinationComponent} from "./SyncDestinationComponent";
 import {CodeSynchronizer} from "../../sync";
 
 /**
@@ -42,7 +42,7 @@ export class ConfigurationDataProvider
             new BundleTargetComponent(this.configModel),
             new AuthTypeComponent(this.connectionManager, this.configModel),
             new ClusterComponent(this.connectionManager, this.configModel),
-            new WorkspaceComponent(
+            new SyncDestinationComponent(
                 this.codeSynchronizer,
                 this.connectionManager,
                 this.configModel