+++ packages/databricks-vscode/src/ui/wizard.ts
@@ -50,6 +50,7 @@ interface QuickAutoCompleteParameters {
     buttons?: QuickInputButton[];
     shouldResume: () => Thenable<boolean>;
     items: Array<QuickPickItem>;
+    ignoreFocusOut: boolean;
 }
 
 interface QuickPickParameters<T extends QuickPickItem> {
@@ -61,6 +62,7 @@ interface QuickPickParameters<T extends QuickPickItem> {
     placeholder: string;
     buttons?: QuickInputButton[];
     shouldResume: () => Thenable<boolean>;
+    ignoreFocusOut: boolean;
 }
 
 interface InputBoxParameters {
@@ -127,6 +129,7 @@ export class MultiStepInput {
         buttons,
         shouldResume,
         items,
+        ignoreFocusOut,
     }: QuickAutoCompleteParameters): Promise<
         | string
         | (QuickAutoCompleteParameters extends {buttons: (infer I)[]}
@@ -149,6 +152,7 @@ export class MultiStepInput {
                 input.totalSteps = totalSteps;
                 input.placeholder = placeholder;
                 input.items = [...items];
+                input.ignoreFocusOut = ignoreFocusOut;
 
                 disposables.push(
                     input.onDidChangeValue(async () => {
@@ -251,6 +255,7 @@ export class MultiStepInput {
         placeholder,
         buttons,
         shouldResume,
+        ignoreFocusOut,
     }: P) {
         const disposables: Disposable[] = [];
         try {
@@ -263,6 +268,7 @@ export class MultiStepInput {
                 input.totalSteps = totalSteps;
                 input.placeholder = placeholder;
                 input.items = items;
+                input.ignoreFocusOut = ignoreFocusOut;
                 if (activeItem) {
                     input.activeItems = [activeItem];
                 }